move code to init
